+++ spake2plus/cli/constants.py
@@ -1,4 +1,4 @@
-from spake2plus.ciphersuites import (
+from spake2plus.ciphersuites.ciphersuites import (
     CiphersuiteP256_SHA256,
     CiphersuiteP256_SHA512,
     CiphersuiteP384_SHA256,
