+++ spake2plus/exceptions/exceptions.py
