+++ spake2plus/roles/prover.py
@@ -1,8 +1,8 @@
 import math
-from spake2plus.exceptions import InvalidInputError
-from spake2plus.parameters import Parameters
-from spake2plus.role import Role
-from spake2plus.utils import (
+from spake2plus.exceptions.exceptions import InvalidInputError
+from spake2plus.protocol.parameters import Parameters
+from spake2plus.roles.role import Role
+from spake2plus.utils.utils import (
     encode_point_uncompressed,
     decode_point_uncompressed,
     get_len,
