+++ spake2plus/utils/utils.py
